Remove unused imports and casts

+++ src/main/java/com/jusipat/castleblocks/block/PitchFluid.java
@@ -115,7 +115,7 @@ public abstract class PitchFluid extends FlowableFluid {
 		}
 
 		public int getLevel(FluidState state) {
-			return (Integer)state.get(LEVEL);
+			return state.get(LEVEL);
 		}
 
 		public boolean isStill(FluidState state) {
